cicd: Add arm cross-compile test.

+++ examples/saml20/smtp-server-saml20.c
@@ -76,8 +76,7 @@ callback (Gsasl * ctx, Gsasl_session * sctx, Gsasl_property prop)
 		|| (idp[n] >= 'A' && idp[n] <= 'Z')
 		|| (idp[n] >= '0' && idp[n] <= '9') || idp[n] == '.'))
 	    {
-	      printf ("Cannot handle identifier (%zu): %s\n",
-		      (unsigned long) n, idp);
+	      printf ("Cannot handle identifier (%zu): %s\n", n, idp);
 	      return GSASL_AUTHENTICATION_ERROR;
 	    }
